axios and cors fixes and minor changes
