+++ .gitignore
@@ -1,0 +1,5 @@
+# Environment configuration
+**/.env
+
+# VS Code Configuration
+**/.vscode/
