+++ frontend/lib/axios.ts
@@ -1,0 +1,9 @@
+import axios from "axios";
+
+const options = {
+  baseURL: process.env.NEXT_PUBLIC_API_URL,
+  withCredentials: true,
+};
+
+export const API = axios.create(options);
+
